Donor adjustment candidates

Provides donor adjustment candidates, with associated descriptors, in order of the proximity to the centroid of the subject catchment.

Usage

DonAdj(CDs, N = 10, UrbMax = 1)

Arguments

CDs

catchment descriptors derived from either GetCDs or CDsXML for the site of interest

N

number of sites provided; default is 10

UrbMax

a maximum value for URBEXT (the default is 0.03). Any sites with UBEXT2015 above UrbMax will not be included in the results.

Value

A data.frame with rownames of site references and columns of catchment descriptors and distance from subject site.

Details

This function provides the closest N catchments for consideration for QMED donor adjustment.

Author

Anthony Hammond

Examples

# Get some CDs and output candidate donor sites
cds_54022 <- GetCDs(54022)
DonAdj(cds_54022)
#>          AREA ALTBAR ASPBAR ASPVAR BFIHOST19scaled BFIHOST19 BFIHOST DPLBAR
#> 54022   8.748    501    116   0.43           0.266     0.266   0.323   2.94
#> 54092   3.188    514     96   0.47           0.265     0.265   0.329   2.18
#> 54091   3.458    529    131   0.40           0.266     0.266   0.304   2.25
#> 55010  27.250    482    137   0.23           0.303     0.304   0.386   4.98
#> 54080 174.400    349     75   0.13           0.391     0.406   0.448  15.13
#> 63002 181.900    341    259   0.14           0.358     0.374   0.435  27.44
#> 64006  47.020    262    264   0.24           0.420     0.424   0.504   9.22
#> 55005 164.500    392    118   0.10           0.360     0.361   0.419  18.22
#> 55026 172.100    387    125   0.10           0.366     0.366   0.423  18.19
#> 54025  53.330    342     58   0.10           0.397     0.396   0.439   7.78
#>       DPSBAR FARL2015  FARL  FPEXT   LDP PROPWET RMED.1H RMED.1D RMED.2D
#> 54022  180.3   1.0000 1.000 0.0098  5.78    0.66    13.2    72.9    97.8
#> 54092  214.3   1.0000 1.000 0.0015  4.08    0.66    13.2    73.3    98.2
#> 54091  159.4   1.0000 1.000 0.0118  4.43    0.66    13.4    72.9    98.0
#> 55010  211.7   0.9692 1.000 0.0188  9.99    0.66    13.0    72.3    96.8
#> 54080  166.7   0.9279 0.928 0.0294 27.25    0.64    11.1    52.5    68.9
#> 63002  194.0   0.8875 0.898 0.0537 46.15    0.66    11.0    52.3    69.3
#> 64006  204.2   0.9833 0.983 0.0196 17.60    0.66    10.0    42.8    57.3
#> 55005  183.4   0.9922 0.997 0.0414 36.96    0.59    11.1    53.1    70.7
#> 55026  180.2   0.9926 0.997 0.0409 37.65    0.59    11.1    52.7    70.0
#> 54025  161.3   1.0000 1.000 0.0241 15.22    0.59    10.1    42.2    55.5
#>       SAAR9120 SAAR6190 SAAR4170 SPRHOST URBEXT2015 URBEXT2000 URBEXT1990
#> 54022     2677     2481     2254   52.68     0.0000     0.0000     0.0000
#> 54092     2704     2536     2323   52.22     0.0000     0.0000     0.0000
#> 54091     2716     2513     2227   54.11     0.0000     0.0000     0.0000
#> 55010     2481     2341     2320   47.73     0.0004     0.0000     0.0001
#> 54080     1778     1611     1548   40.92     0.0050     0.0026     0.0014
#> 63002     1907     1756     1769   43.99     0.0056     0.0012     0.0007
#> 64006     1644     1496     1508   38.47     0.0045     0.0014     0.0012
#> 55005     1795     1656     1621   43.69     0.0023     0.0014     0.0008
#> 55026     1772     1636     1603   43.36     0.0034     0.0025     0.0015
#> 54025     1431     1268     1291   40.26     0.0016     0.0007     0.0000
#>       DrainDens  CEast CNorth   Lcv   LSkew  LKurt     L1     L2  N Suitability
#> 54022     1.260 283260 288067 0.156  0.1310 0.2870  14.80  2.300 40     Pooling
#> 54092     1.153 282914 287310 0.112 -0.0129 0.0952   6.28  0.704 51        QMED
#> 54091     1.171 283142 288978 0.138  0.2650 0.4310   6.14  0.849 49        QMED
#> 55010     1.164 281567 284221 0.168  0.1760 0.1640  54.00  9.100 54        QMED
#> 54080     1.236 291590 284733 0.243  0.2760 0.2150 109.00 26.600 24     Pooling
#> 63002     1.282 274413 283210 0.169  0.1730 0.2990  97.30 16.400 23        QMED
#> 64006     1.369 269502 288174 0.263  0.3580 0.2960  22.00  5.790 52        QMED
#> 55005     1.209 291754 277165 0.267  0.2230 0.0847 137.00 36.500 31        QMED
#> 55026     1.207 292073 276806 0.167  0.1190 0.1030 124.00 20.600 55     Pooling
#> 54025     1.157 296850 279005 0.196  0.2590 0.2150  26.20  5.120 55     Pooling
#>         QMED QMEDcd   Distance
#> 54022  14.70  17.70  0.0000000
#> 54092   6.33   7.58  0.8323251
#> 54091   5.99   8.14  0.9186104
#> 55010  50.40  37.50  4.2021381
#> 54080  93.70  98.40  8.9724275
#> 63002  94.50 104.00 10.0925645
#> 64006  18.60  32.80 13.7584161
#> 55005 115.00 125.00 13.8203343
#> 55026 116.00 127.00 14.2996185
#> 54025  23.50  35.00 16.3342568