Get an annual maximum sample from the National River Flow Archive sites suitable for pooling

Extracts the annual maximum peak flow sample and associated dates for the site of interest.

Usage

GetAM(ref)

Arguments

ref

the site reference of interest (numeric)

Value

A data.frame with columns; Date, Flow, and id

Author

Anthony Hammond

Examples

# Get an AMAX sample and display it in the console
GetAM(203018)
#>          Date    Flow     id
#> 1  1971-04-25  82.241 203018
#> 2  1972-04-01  71.814 203018
#> 3  1973-01-19  55.235 203018
#> 4  1973-12-20  70.028 203018
#> 5  1974-12-29  46.774 203018
#> 6  1976-01-02  46.109 203018
#> 7  1976-10-23  98.876 203018
#> 8  1978-01-28  58.536 203018
#> 9  1978-12-28 117.368 203018
#> 10 1979-11-14  70.302 203018
#> 11 1980-10-23  82.136 203018
#> 12 1981-10-03 119.691 203018
#> 13 1982-11-08  58.527 203018
#> 14 1984-02-21  85.928 203018
#> 15 1985-09-08  93.412 203018
#> 16 1986-08-26 115.463 203018
#> 17 1987-08-17  54.557 203018
#> 18 1987-10-22 163.842 203018
#> 19 1988-10-26  79.716 203018
#> 20 1990-01-25  92.876 203018
#> 21 1990-11-24  96.991 203018
#> 22 1991-12-23  70.242 203018
#> 23 1993-05-31  74.646 203018
#> 24 1994-02-27  93.479 203018
#> 25 1995-01-21  64.640 203018
#> 26 1995-11-16 113.945 203018
#> 27 1996-12-04  62.771 203018
#> 28 1997-11-26 106.749 203018
#> 29 1999-09-20  98.452 203018
#> 30 1999-12-25 104.855 203018
#> 31 2000-10-10  88.168 203018
#> 32 2001-12-05  81.944 203018
#> 33 2002-10-22  88.759 203018
#> 34 2003-11-29  66.825 203018
#> 35 2005-01-08  56.781 203018
#> 36 2005-10-24  80.859 203018
#> 37 2007-08-12  62.754 203018
#> 38 2008-08-17 212.674 203018
#> 39 2009-01-31 119.542 203018
#> 40 2010-03-31  79.067 203018
#> 41 2010-12-27  64.231 203018
#> 42 2012-06-23 126.939 203018
#> 43 2013-01-26  89.543 203018
#> 44 2014-02-14  81.404 203018
#> 45 2014-12-23  85.301 203018
#> 46 2015-12-30 104.424 203018
#> 47 2017-03-04  57.081 203018
#> 48 2018-02-10  62.222 203018
#> 49 2018-11-07  74.872 203018
#> 50 2020-02-13  66.034 203018
#> 51 2021-02-03  74.424 203018
#> 52 2021-12-26  53.341 203018
#> 53 2023-08-19  63.757 203018
#> 54 2023-11-01 116.695 203018

# Save an AMAX sample as an object
am_203018 <- GetAM(203018)