Low Flows

A function to estimate lower flow quantiles in ungauged catchments.

Usage

LowFlows(
  CDs = NULL,
  AREA = NULL,
  SAAR = NULL,
  BFIHOST = NULL,
  Exclude = NULL,
  FARLRange = NULL
)

Arguments

CDs

Catchment descriptors derived from the GetCDs or CDsXML function.

AREA

Catchment area (km2) - for when CDs is not applied

SAAR

Average annual rainfall (mm) - for when CDs is not applied

BFIHOST

An estimate of baseflow index - for when CDs is not applied

Exclude

A site reference. This is to exclude sites that you do not want used in the estimate. For example, if you're seeing how the function performs on a gauged site, you may want to exclude it from the analysis.

FARLRange

A vector of length 2. For example c(0.9,1). This determines a FARL range for the catchments you wish to be included in the analysis. Primarily this is to exclude sites which have significant reservoir or lake influence when the site of interest does not. If it is NULL (default) all NRFA sites are included.

Value

A list. The first element of which is a data.frame with one column of flow estimates. The row names denote the name of each estimate. The second element is the dataframe of catchments used in the analysis, with the relevant descriptors and weighting.

Details

This function provides estimates of the mean flow, Q95, Q70, Q50, Q10, and Q5. The function works by finding the 30 catchments in the NRFA data set with the most similar SAAR9120 to the subject site. The observed flows for those catchments are scaled by the catchment area. Then a weighted average is taken and multiplied by the subject site catchment area for the final estimate. The weighting is done by Euclidean distance based on SAAR9120 and BFIHOST19scaled. These are weighted based on the "Spearman's Rho" correlation coefficient of these descriptors to the scaled mean flows.

Author

Anthony Hammond

Examples

# Get some catchment descriptors, then estimate the flows
if (FALSE) { # \dontrun{
CDs_27083 <- GetCDs(27083)
LowFlows(CDs_27083)
} # }
# Now estimate again but remove gauge 27083 from the analysis
if (FALSE) { # \dontrun{
LowFlows(CDs_27083, Exclude = 27083)
} # }